Reimplement using CRTP/inline classes to avoid linker challenges
