+++ onnxruntime/core/providers/shared_library/provider_api.h
@@ -261,10 +261,6 @@ std::string GetEnvironmentVar(const std::string& var_name);
 
 namespace profiling {
 
-  class GPUTracerManager;
-  class ProfilerActivityBuffer;
-  class GPUProfilerBase;
-
   std::string demangle(const char* name);
   std::string demangle(const std::string& name);
